・店舗のモデル構成を修正する。緯度経度の型をDoubleとする。距離のプロパティを追加する。 ・店舗情報取得の際に、現在地の緯度経度と店舗の緯度経度から距離を計算する。 ・店舗画面に、店舗一覧のTableViewを追加する。

+++ IOS_Xacdonald/ViewModel/ShopViewModel.swift
@@ -11,7 +11,7 @@ import RxSwift
 
 class ShopViewModel {
     
-    func getShop(lat: Float, lon: Float, query: String) -> Observable<[Shop]> {
+    func getShop(lat: Double, lon: Double, query: String) -> Observable<[Shop]> {
         let subject = PublishSubject<[Shop]>()
         let observable = subject.asObserver()
         ShopRepository.getShop(lat: lat, lon: lon, query: query)
